locale: rename keymap-util.[ch] -> localed-util.[ch]
As the file contains not only keymap related functions.

(cherry picked from commit 3e5203b3eb2b06ef346eef13d64430542bbfb5c4)

Related: #2087652
