+++ src/locale/localed-util.c
@@ -15,8 +15,7 @@
 #include "fileio.h"
 #include "fs-util.h"
 #include "kbd-util.h"
-#include "keymap-util.h"
-#include "locale-util.h"
+#include "localed-util.h"
 #include "macro.h"
 #include "mkdir-label.h"
 #include "nulstr-util.h"
